LayoutTests/media/caption-display-settings/caption-display-settings-expected.txt


EXPECTED (video.showCaptionDisplaySettings != 'undefined') OK
END OF TEST
